MAINT: Move _xobj_to_image
In its own file is a more consistent location.
